test code (미완성)
